add cap info to form
