+++ app/views/spotlight/resources/upload/_form.html.erb
@@ -1,0 +1,22 @@
+<%= bootstrap_form_for([current_exhibit, @resource.becomes(Spotlight::Resources::Upload)], layout: :horizontal, label_col: 'col-md-2', control_col: 'col-sm-6 col-md-6', html: { class: 'item-upload-form', multipart: true } ) do |f| %>
+  <p> <%= t('cap-notice', size: format_max_total(current_exhibit.id), max: format_max_bytes) %></p>
+  <%= f.url_field :url, type: "file", help: t('.url-field.help', extensions: Spotlight::Engine.config.allowed_upload_extensions.join(' ')), label: "File" %>
+  <%= f.fields_for :data do |d| %>
+    <% Spotlight::Resources::Upload.fields(current_exhibit).each do |config| %>
+      <%= d.send(config.form_field_type, config.field_name, label: uploaded_field_label(config)) %>
+    <% end %>
+
+    <% current_exhibit.custom_fields.each do |custom_field| %>
+      <%= render partial: "spotlight/custom_fields/form_group/#{custom_field.field_type}", locals: { inline: true, f: d, field: custom_field, value: nil }  %>
+    <% end %>
+  <% end %>
+  <div class="form-actions">
+    <div class="primary-actions">
+      <%= hidden_field_tag :tab, 'upload' %>
+      <%= cancel_link @resource, :back, class: 'btn-sizing' %>
+      <%= f.submit t('.add_item_and_continue'), name: 'add-and-continue', class: 'btn btn-secondary' %>
+      <%= f.submit t('.add_item'), class: 'btn btn-primary' %>
+    </div>
+  </div>
+<% end %>
+
